requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1
aiogram==3.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.26.0
loguru==0.7.2